// src/app/interfaces/project-metadata.ts
export interface ProjectMetadata {
  totalProjects: number;
  averageProjectBudget: number;
  totalProjectsBudget: number;
  averageProjectsPerProjectOwner: number;
  numProjectOwners: number;
  projectsCreatedInLastMonth: number;
  projectsCreatedInLastYear: number;
  projectsModifiedInLastMonth: number;
  projectsModifiedInLastYear: number;
}